strategy

+++ src/main/java/br/ufscar/dc/dsw/AA2/services/StrategyService.java
@@ -1,16 +1,19 @@
 package br.ufscar.dc.dsw.AA2.services;
 
-import br.ufscar.dc.dsw.AA2.dtos.StrategyDTO;
+import br.ufscar.dc.dsw.AA2.dtos.strategy.StrategyCreateDTO;
+import br.ufscar.dc.dsw.AA2.mappers.StrategyMapper;
 import br.ufscar.dc.dsw.AA2.models.Image;
 import br.ufscar.dc.dsw.AA2.models.Strategy;
 import br.ufscar.dc.dsw.AA2.repositories.ImageRepository;
 import br.ufscar.dc.dsw.AA2.repositories.StrategyRepository;
 import br.ufscar.dc.dsw.AA2.storage.StorageProvider;
+import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.util.List;
+import java.util.Optional;
 import java.util.UUID;
 
 @Service
@@ -24,35 +27,69 @@ public class StrategyService {
     @Autowired
     private StorageProvider storageProvider;
 
+    @Autowired
+    private StrategyMapper strategyMapper;
+
+    @Transactional()
     public List<Strategy> getAll() {
         return strategyRepository.findAll();
     }
 
-    public void deleteById(UUID id) {
-        strategyRepository.deleteById(id);
+    @Transactional()
+    public Optional<Strategy> getById(UUID id) {
+        return strategyRepository.findById(id);
     }
 
-    public void insert(StrategyDTO strategyDTO, List<MultipartFile> imagesFiles) {
+    public Strategy create(StrategyCreateDTO strategyDTO, List<MultipartFile> imagesFiles) {
+
         Strategy strategy = new Strategy();
+
+        strategy.setName(strategyDTO.name());
+        strategy.setDescription(strategyDTO.description());
+        strategy.setExamples(strategyDTO.examples());
+        strategy.setTips(strategyDTO.tips());
+
+        strategyRepository.save(strategy);
+
+        if (imagesFiles != null && !imagesFiles.isEmpty()) {
+            this.uploadImages(strategy, imagesFiles);
+        }
+
+        return strategy;
+    }
+
+    public Strategy update(UUID id, StrategyCreateDTO strategyDTO) {
+        Strategy strategy = strategyRepository.findById(id)
+                .orElseThrow(() -> new RuntimeException("Strategy not found with id: " + id));
+
         strategy.setName(strategyDTO.name());
         strategy.setDescription(strategyDTO.description());
         strategy.setTips(strategyDTO.tips());
         strategy.setExamples(strategyDTO.examples());
 
-        strategyRepository.save(strategy);
+        return strategyRepository.save(strategy);
+    }
 
-        this.uploadImages(strategy, imagesFiles);
+
+    public void deleteById(UUID id) {
+        if (!strategyRepository.existsById(id)) {
+            throw new RuntimeException("Strategy not found with id: " + id);
+        }
+
+        strategyRepository.deleteById(id);
     }
 
     private void uploadImages(Strategy strategy, List<MultipartFile> imagesFiles) {
         imagesFiles.forEach(imageFile -> {
-            storageProvider.validateFile(imageFile);
-            String imageUrl = storageProvider.store(imageFile);
+            if(imageFile != null && !imageFile.isEmpty()) {
+                storageProvider.validateFile(imageFile);
+                String imageUrl = storageProvider.store(imageFile);
 
-            Image image = new Image();
-            image.setUrl(imageUrl);
-            image.setStrategy(strategy);
-            imageRepository.save(image);
+                Image image = new Image();
+                image.setUrl(imageUrl);
+                image.setStrategy(strategy);
+                imageRepository.save(image);
+            }
         });
     }
 }
